Update Flask app - add gethostname function
The Flask app has been updated to include a new function called gethostname. This function returns the hostname of the machine on which the app is running. The function utilizes the `uname` method from the `os` module to fetch the hostname. This addition allows the app to display the hostname in the HTML response when accessing the root route ("/").